refactor: separate chat to its own component
